+++ index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-//services exposed to outside
+// services exposed to outside
 module.exports = require('./lib/setup');
 module.exports.db = require('./lib/services/db');
 module.exports.composer = require('./lib/services/composer');
@@ -9,3 +9,4 @@ module.exports.schedule = require('./lib/services/schedule');
 module.exports.pages = require('./lib/services/pages');
 module.exports.sites = require('./lib/services/sites');
 module.exports.references = require('./lib/services/references');
+module.exports.log = require('./lib/services/log');
