+++ lib/bootstrap.js
@@ -7,7 +7,7 @@ const path = require('path'),
   siteService = require('./services/sites'),
   references = require('./services/references'),
   db = require('./services/db'),
-  log = require('./log').withStandardPrefix(__filename),
+  log = require('./services/log').withStandardPrefix(__filename),
   chalk = require('chalk');
 
 /**
