+++ lib/control.js
@@ -2,7 +2,7 @@
 
 let memoryLeakThreshold = 1000;
 const _ = require('lodash'),
-  log = require('./log').withStandardPrefix(__filename),
+  log = require('./services/log').withStandardPrefix(__filename),
   minute = 60000;
 
 /**
