+++ lib/html-composer.js
@@ -11,7 +11,7 @@ const _ = require('lodash'),
   components = require('./services/components'),
   control = require('./control'),
   db = require('./services/db'),
-  log = require('./log').withStandardPrefix(__filename),
+  log = require('./services/log').withStandardPrefix(__filename),
   multiplexTemplates = require('multiplex-templates'),
   references = require('./services/references'),
   composer = require('./services/composer'),
