+++ lib/responses.js
@@ -9,7 +9,7 @@
 const _ = require('lodash'),
   db = require('./services/db'),
   bluebird = require('bluebird'),
-  log = require('./log').withStandardPrefix(__filename),
+  log = require('./services/log').withStandardPrefix(__filename),
   filter = require('through2-filter'),
   referencedProperty = '_ref';
