+++ lib/services/components.js
@@ -14,7 +14,7 @@ const _ = require('lodash'),
   uid = require('../uid'),
   files = require('../files'),
   timer = require('../timer'),
-  log = require('../log').withStandardPrefix(__filename),
+  log = require('./log').withStandardPrefix(__filename),
   schema = require('../schema'),
   siteService = require('./sites'),
   references = require('./references'),
