+++ lib/services/notifications.js
@@ -3,7 +3,7 @@
 const _ = require('lodash'),
   bluebird = require('bluebird'),
   contentType = 'Content-type',
-  log = require('../log').withStandardPrefix(__filename),
+  log = require('./log').withStandardPrefix(__filename),
   rest = require('../rest');
 
 /**
