+++ lib/services/pages.js
@@ -3,12 +3,12 @@
 let timeoutConstant = 4000;
 const _ = require('lodash'),
   bluebird = require('bluebird'),
-  components = require('../services/components'),
+  components = require('./components'),
   composer = require('./composer'),
-  db = require('../services/db'),
-  log = require('../log').withStandardPrefix(__filename),
+  db = require('./db'),
+  log = require('./log').withStandardPrefix(__filename),
   notifications = require('./notifications'),
-  references = require('../services/references'),
+  references = require('./references'),
   siteService = require('./sites'),
   timer = require('../timer'),
   uid = require('../uid'),
