+++ lib/services/schedule.js
@@ -4,12 +4,12 @@ let interval,
   intervalDelay = 50000 + Math.floor(Math.random() * 10000); // just under one minute
 const _ = require('lodash'),
   bluebird = require('bluebird'),
-  db = require('../services/db'),
-  log = require('../log').withStandardPrefix(__filename),
+  db = require('./db'),
+  log = require('./log').withStandardPrefix(__filename),
   publishProperty = 'publish',
-  references = require('../services/references'),
+  references = require('./references'),
   rest = require('../rest'),
-  siteService = require('../services/sites'),
+  siteService = require('./sites'),
   scheduledAtProperty = 'at',
   scheduledVersion = 'scheduled',
   uid = require('../uid');
