+++ lib/streams/sitemap-text-transform.js
@@ -6,7 +6,7 @@
 
 'use strict';
 
-const log = require('../log').withStandardPrefix(__filename),
+const log = require('../services/log').withStandardPrefix(__filename),
   util = require('util'),
   Transform = require('stream').Transform;
