+++ test/fixtures/api-accepts.js
@@ -9,7 +9,7 @@ const _ = require('lodash'),
   db = require('../../lib/services/db'),
   bluebird = require('bluebird'),
   multiplex = require('multiplex-templates'),
-  log = require('../../lib/log'),
+  log = require('../../lib/services/log'),
   schema = require('../../lib/schema'),
   siteService = require('../../lib/services/sites'),
   expect = require('chai').expect,
